+++ resources/views/authorize.blade.php
@@ -65,7 +65,7 @@
                         <div class="buttons">
                             <!-- Authorize Button -->
                             <form method="post" action="{{ route('passport.authorizations.approve') }}">
-                                {{ csrf_field() }}
+                                @csrf
 
                                 <input type="hidden" name="state" value="{{ $request->state }}">
                                 <input type="hidden" name="client_id" value="{{ $client->id }}">
@@ -74,8 +74,8 @@
 
                             <!-- Cancel Button -->
                             <form method="post" action="{{ route('passport.authorizations.deny') }}">
-                                {{ csrf_field() }}
-                                {{ method_field('DELETE') }}
+                                @csrf
+                                @method('DELETE')
 
                                 <input type="hidden" name="state" value="{{ $request->state }}">
                                 <input type="hidden" name="client_id" value="{{ $client->id }}">