+++ src/Bridge/RefreshTokenRepository.php
@@ -2,20 +2,20 @@
 
 namespace Laravel\Passport\Bridge;
 
-use Illuminate\Database\Connection;
 use Illuminate\Contracts\Events\Dispatcher;
 use Laravel\Passport\Events\RefreshTokenCreated;
 use League\OAuth2\Server\Entities\RefreshTokenEntityInterface;
 use League\OAuth2\Server\Repositories\RefreshTokenRepositoryInterface;
+use Laravel\Passport\RefreshTokenRepository as PassportRefreshTokenRepository;
 
 class RefreshTokenRepository implements RefreshTokenRepositoryInterface
 {
     /**
-     * The database connection.
+     * The refresh token repository instance.
      *
      * @var \Illuminate\Database\Connection
      */
-    protected $database;
+    protected $refreshTokenRepository;
 
     /**
      * The event dispatcher instance.
@@ -27,14 +27,14 @@ class RefreshTokenRepository implements RefreshTokenRepositoryInterface
     /**
      * Create a new repository instance.
      *
-     * @param  \Illuminate\Database\Connection  $database
+     * @param  \Laravel\Passport\RefreshTokenRepository  $refreshTokenRepository
      * @param  \Illuminate\Contracts\Events\Dispatcher  $events
      * @return void
      */
-    public function __construct(Connection $database, Dispatcher $events)
+    public function __construct(PassportRefreshTokenRepository $refreshTokenRepository, Dispatcher $events)
     {
         $this->events = $events;
-        $this->database = $database;
+        $this->refreshTokenRepository = $refreshTokenRepository;
     }
 
     /**
@@ -50,7 +50,7 @@ class RefreshTokenRepository implements RefreshTokenRepositoryInterface
      */
     public function persistNewRefreshToken(RefreshTokenEntityInterface $refreshTokenEntity)
     {
-        $this->database->table('oauth_refresh_tokens')->insert([
+        $this->refreshTokenRepository->create([
             'id' => $id = $refreshTokenEntity->getIdentifier(),
             'access_token_id' => $accessTokenId = $refreshTokenEntity->getAccessToken()->getIdentifier(),
             'revoked' => false,
@@ -65,8 +65,7 @@ class RefreshTokenRepository implements RefreshTokenRepositoryInterface
      */
     public function revokeRefreshToken($tokenId)
     {
-        $this->database->table('oauth_refresh_tokens')
-                    ->where('id', $tokenId)->update(['revoked' => true]);
+        $this->refreshTokenRepository->revokeRefreshToken($tokenId);
     }
 
     /**
@@ -74,9 +73,6 @@ class RefreshTokenRepository implements RefreshTokenRepositoryInterface
      */
     public function isRefreshTokenRevoked($tokenId)
     {
-        $refreshToken = $this->database->table('oauth_refresh_tokens')
-                    ->where('id', $tokenId)->first();
-
-        return $refreshToken === null || $refreshToken->revoked;
+        return $this->refreshTokenRepository->isRefreshTokenRevoked($tokenId);
     }
 }