+++ src/Http/Controllers/AuthorizationController.php
@@ -14,7 +14,7 @@ use Illuminate\Contracts\Routing\ResponseFactory;
 
 class AuthorizationController
 {
-    use HandlesOAuthErrors;
+    use ConvertsPsrResponses;
 
     /**
      * The authorization server.
@@ -57,30 +57,28 @@ class AuthorizationController
                               ClientRepository $clients,
                               TokenRepository $tokens)
     {
-        return $this->withErrorHandling(function () use ($psrRequest, $request, $clients, $tokens) {
-            $authRequest = $this->server->validateAuthorizationRequest($psrRequest);
+        $authRequest = $this->server->validateAuthorizationRequest($psrRequest);
 
-            $scopes = $this->parseScopes($authRequest);
+        $scopes = $this->parseScopes($authRequest);
 
-            $token = $tokens->findValidToken(
-                $user = $request->user(),
-                $client = $clients->find($authRequest->getClient()->getIdentifier())
-            );
+        $token = $tokens->findValidToken(
+            $user = $request->user(),
+            $client = $clients->find($authRequest->getClient()->getIdentifier())
+        );
 
-            if (($token && $token->scopes === collect($scopes)->pluck('id')->all()) ||
-                $client->skipsAuthorization()) {
-                return $this->approveRequest($authRequest, $user);
-            }
+        if (($token && $token->scopes === collect($scopes)->pluck('id')->all()) ||
+            $client->skipsAuthorization()) {
+            return $this->approveRequest($authRequest, $user);
+        }
 
-            $request->session()->put('authRequest', $authRequest);
+        $request->session()->put('authRequest', $authRequest);
 
-            return $this->response->view('passport::authorize', [
-                'client' => $client,
-                'user' => $user,
-                'scopes' => $scopes,
-                'request' => $request,
-            ]);
-        });
+        return $this->response->view('passport::authorize', [
+            'client' => $client,
+            'user' => $user,
+            'scopes' => $scopes,
+            'request' => $request,
+        ]);
     }
 
     /**