+++ src/Http/Middleware/CheckClientCredentialsForAnyScope.php
@@ -3,11 +3,16 @@
 namespace Laravel\Passport\Http\Middleware;
 
 use Closure;
+use Zend\Diactoros\StreamFactory;
+use Zend\Diactoros\ResponseFactory;
+use Laravel\Passport\TokenRepository;
+use Zend\Diactoros\UploadedFileFactory;
 use League\OAuth2\Server\ResourceServer;
+use Zend\Diactoros\ServerRequestFactory;
 use Illuminate\Auth\AuthenticationException;
 use Laravel\Passport\Exceptions\MissingScopeException;
 use League\OAuth2\Server\Exception\OAuthServerException;
-use Symfony\Bridge\PsrHttpMessage\Factory\DiactorosFactory;
+use Symfony\Bridge\PsrHttpMessage\Factory\PsrHttpFactory;
 
 class CheckClientCredentialsForAnyScope
 {
@@ -19,14 +24,23 @@ class CheckClientCredentialsForAnyScope
     protected $server;
 
     /**
+     * Token Repository.
+     *
+     * @var \Laravel\Passport\TokenRepository
+     */
+    protected $repository;
+
+    /**
      * Create a new middleware instance.
      *
      * @param  \League\OAuth2\Server\ResourceServer  $server
+     * @param  \Laravel\Passport\TokenRepository  $repository
      * @return void
      */
-    public function __construct(ResourceServer $server)
+    public function __construct(ResourceServer $server, TokenRepository $repository)
     {
         $this->server = $server;
+        $this->repository = $repository;
     }
 
     /**
@@ -40,7 +54,12 @@ class CheckClientCredentialsForAnyScope
      */
     public function handle($request, Closure $next, ...$scopes)
     {
-        $psr = (new DiactorosFactory)->createRequest($request);
+        $psr = (new PsrHttpFactory(
+            new ServerRequestFactory,
+            new StreamFactory,
+            new UploadedFileFactory,
+            new ResponseFactory
+        ))->createRequest($request);
 
         try {
             $psr = $this->server->validateAuthenticatedRequest($psr);
@@ -48,7 +67,7 @@ class CheckClientCredentialsForAnyScope
             throw new AuthenticationException;
         }
 
-        if ($this->validateScopes($psr, $scopes)) {
+        if ($this->validate($psr, $scopes)) {
             return $next($request);
         }
 
@@ -56,20 +75,27 @@ class CheckClientCredentialsForAnyScope
     }
 
     /**
-     * Validate the scopes on the incoming request.
+     * Validate the scopes and token on the incoming request.
      *
      * @param  \Psr\Http\Message\ServerRequestInterface $psr
      * @param  array  $scopes
      * @return bool
+     * @throws \Illuminate\Auth\AuthenticationException
      */
-    protected function validateScopes($psr, $scopes)
+    protected function validate($psr, $scopes)
     {
-        if (in_array('*', $tokenScopes = $psr->getAttribute('oauth_scopes'))) {
+        $token = $this->repository->find($psr->getAttribute('oauth_access_token_id'));
+
+        if (! $token || $token->client->firstParty()) {
+            throw new AuthenticationException;
+        }
+
+        if (in_array('*', $token->scopes)) {
             return true;
         }
 
         foreach ($scopes as $scope) {
-            if (in_array($scope, $tokenScopes)) {
+            if ($token->can($scope)) {
                 return true;
             }
         }