+++ tests/BridgeAccessTokenRepositoryTest.php
@@ -3,7 +3,7 @@
 namespace Laravel\Passport\Tests;
 
 use Mockery as m;
-use Carbon\Carbon;
+use Carbon\CarbonImmutable;
 use PHPUnit\Framework\TestCase;
 use Laravel\Passport\Bridge\Scope;
 use Laravel\Passport\Bridge\Client;
@@ -21,7 +21,7 @@ class BridgeAccessTokenRepositoryTest extends TestCase
 
     public function test_access_tokens_can_be_persisted()
     {
-        $expiration = Carbon::now();
+        $expiration = CarbonImmutable::now();
 
         $tokenRepository = m::mock(TokenRepository::class);
         $events = m::mock(Dispatcher::class);
@@ -39,13 +39,29 @@ class BridgeAccessTokenRepositoryTest extends TestCase
 
         $events->shouldReceive('dispatch')->once();
 
-        $accessToken = new AccessToken(2, [new Scope('scopes')]);
+        $accessToken = new AccessToken(2, [new Scope('scopes')], new Client('client-id', 'name', 'redirect'));
         $accessToken->setIdentifier(1);
         $accessToken->setExpiryDateTime($expiration);
-        $accessToken->setClient(new Client('client-id', 'name', 'redirect'));
 
         $repository = new AccessTokenRepository($tokenRepository, $events);
 
         $repository->persistNewAccessToken($accessToken);
+    }
+
+    public function test_can_get_new_access_token()
+    {
+        $tokenRepository = m::mock(TokenRepository::class);
+        $events = m::mock(Dispatcher::class);
+        $repository = new AccessTokenRepository($tokenRepository, $events);
+        $client = new Client('client-id', 'name', 'redirect');
+        $scopes = [new Scope('place-orders'), new Scope('check-status')];
+        $userIdentifier = 123;
+
+        $token = $repository->getNewToken($client, $scopes, $userIdentifier);
+
+        $this->assertInstanceOf(AccessToken::class, $token);
+        $this->assertEquals($client, $token->getClient());
+        $this->assertEquals($scopes, $token->getScopes());
+        $this->assertEquals($userIdentifier, $token->getUserIdentifier());
     }
 }