+++ tests/BridgeClientRepositoryTest.php
@@ -38,25 +38,32 @@ class BridgeClientRepositoryTest extends TestCase
         unset($this->clientModelRepository, $this->repository);
     }
 
-    public function test_can_get_client_for_auth_code_grant()
+    public function test_can_get_client()
     {
-        $client = $this->repository->getClientEntity(1, 'authorization_code', 'secret', true);
+        $client = $this->repository->getClientEntity(1);
 
         $this->assertInstanceOf(Client::class, $client);
-        $this->assertNull($this->repository->getClientEntity(1, 'authorization_code', 'wrong-secret', true));
-        $this->assertNull($this->repository->getClientEntity(1, 'client_credentials', 'wrong-secret', true));
+        $this->assertEquals('1', $client->getIdentifier());
+        $this->assertEquals('Client', $client->getName());
+        $this->assertEquals(['http://localhost'], $client->getRedirectUri());
+        $this->assertTrue($client->isConfidential());
     }
 
-    public function test_can_get_client_for_client_credentials_grant()
+    public function test_can_validate_client_for_auth_code_grant()
+    {
+        $this->assertTrue($this->repository->validateClient(1, 'secret', 'authorization_code'));
+        $this->assertFalse($this->repository->validateClient(1, 'wrong-secret', 'authorization_code'));
+        $this->assertFalse($this->repository->validateClient(1, 'wrong-secret', 'client_credentials'));
+    }
+
+    public function test_can_validate_client_for_client_credentials_grant()
     {
         $client = $this->clientModelRepository->findActive(1);
         $client->personal_access_client = true;
 
-        $this->assertInstanceOf(
-            Client::class,
-            $this->repository->getClientEntity(1, 'client_credentials', 'secret', true)
-        );
-        $this->assertNull($this->repository->getClientEntity(1, 'authorization_code', 'secret', true));
+        $this->assertTrue($this->repository->validateClient(1, 'secret', 'client_credentials'));
+        $this->assertFalse($this->repository->validateClient(1, 'wrong-secret', 'client_credentials'));
+        $this->assertFalse($this->repository->validateClient(1, 'secret', 'authorization_code'));
     }
 
     public function test_password_grant_is_permitted()
@@ -64,17 +71,17 @@ class BridgeClientRepositoryTest extends TestCase
         $client = $this->clientModelRepository->findActive(1);
         $client->password_client = true;
 
-        $this->assertInstanceOf(Client::class, $this->repository->getClientEntity(1, 'password', 'secret'));
+        $this->assertTrue($this->repository->validateClient(1, 'secret', 'password'));
     }
 
     public function test_password_grant_is_prevented()
     {
-        $this->assertNull($this->repository->getClientEntity(1, 'password', 'secret'));
+        $this->assertFalse($this->repository->validateClient(1, 'secret', 'password'));
     }
 
     public function test_authorization_code_grant_is_permitted()
     {
-        $this->assertInstanceOf(Client::class, $this->repository->getClientEntity(1, 'authorization_code', 'secret'));
+        $this->assertTrue($this->repository->validateClient(1, 'secret', 'authorization_code'));
     }
 
     public function test_authorization_code_grant_is_prevented()
@@ -82,7 +89,7 @@ class BridgeClientRepositoryTest extends TestCase
         $client = $this->clientModelRepository->findActive(1);
         $client->password_client = true;
 
-        $this->assertNull($this->repository->getClientEntity(1, 'authorization_code', 'secret'));
+        $this->assertFalse($this->repository->validateClient(1, 'secret', 'authorization_code'));
     }
 
     public function test_personal_access_grant_is_permitted()
@@ -90,17 +97,17 @@ class BridgeClientRepositoryTest extends TestCase
         $client = $this->clientModelRepository->findActive(1);
         $client->personal_access_client = true;
 
-        $this->assertInstanceOf(Client::class, $this->repository->getClientEntity(1, 'personal_access', 'secret'));
+        $this->assertTrue($this->repository->validateClient(1, 'secret', 'personal_access'));
     }
 
     public function test_personal_access_grant_is_prevented()
     {
-        $this->assertNull($this->repository->getClientEntity(1, 'personal_access', 'secret'));
+        $this->assertFalse($this->repository->validateClient(1, 'secret', 'personal_access'));
     }
 
     public function test_client_credentials_grant_is_permitted()
     {
-        $this->assertInstanceOf(Client::class, $this->repository->getClientEntity(1, 'client_credentials', 'secret'));
+        $this->assertTrue($this->repository->validateClient(1, 'secret', 'client_credentials'));
     }
 
     public function test_client_credentials_grant_is_prevented()
@@ -108,7 +115,7 @@ class BridgeClientRepositoryTest extends TestCase
         $client = $this->clientModelRepository->findActive(1);
         $client->secret = null;
 
-        $this->assertNull($this->repository->getClientEntity(1, 'client_credentials', 'secret'));
+        $this->assertFalse($this->repository->validateClient(1, 'secret', 'client_credentials'));
     }
 
     public function test_grant_types_allows_request()
@@ -116,7 +123,7 @@ class BridgeClientRepositoryTest extends TestCase
         $client = $this->clientModelRepository->findActive(1);
         $client->grant_types = ['client_credentials'];
 
-        $this->assertInstanceOf(Client::class, $this->repository->getClientEntity(1, 'client_credentials', 'secret'));
+        $this->assertTrue($this->repository->validateClient(1, 'secret', 'client_credentials'));
     }
 
     public function test_grant_types_disallows_request()
@@ -124,7 +131,7 @@ class BridgeClientRepositoryTest extends TestCase
         $client = $this->clientModelRepository->findActive(1);
         $client->grant_types = ['client_credentials'];
 
-        $this->assertNull($this->repository->getClientEntity(1, 'authorization_code', 'secret'));
+        $this->assertFalse($this->repository->validateClient(1, 'secret', 'authorization_code'));
     }
 }