+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/passport/compare/v7.3.4...7.0)
+## [Unreleased](https://github.com/laravel/passport/compare/v7.3.4...master)
+
+### Added
+- Add ability to customize the `RefreshToken` ([#966](https://github.com/laravel/passport/pull/966))
+
+### Changed
+- Rework HandlesOAuthErrors trait to middleware ([#937](https://github.com/laravel/passport/pull/937))
+- Use diactoros 2.0 and psr-http-factory ([aadf603](https://github.com/laravel/passport/commit/aadf603c1f45cfa4bbf954bfc3abc30cdd572683))
+- Replaced helpers with Blade directives ([#939](https://github.com/laravel/passport/pull/939))
+- Use caret for constraints ([d906804](https://github.com/laravel/passport/commit/d906804c2faccca0333801eccfbf6c2fa5afbaee))
 
 
 ## [v7.3.4 (2019-07-30)](https://github.com/laravel/passport/compare/v7.3.3...v7.3.4)